Adds loader function on loader JS
